// src/lib/valvomatic/valvomatic-types.ts
export interface ValvomaticQualitySummaryItem {
  quality: string;
  rollsCount: number;
  totalRollMtr: number;
  totalNetWt: number;
  avgGsm: number;
  productionPcs: number;
  productionKg: number;
  coverPatchOs: number;
  coverPatchDs: number;
  valvePatch: number;
  totalWasteKg: number;
  totalWastePct: number;
  netProductionKg: number;
}

export interface ValvomaticWastageBreakdown {
  loomWasteKg: number;
  loomWastePct: number;
  lamWasteKg: number;
  lamWastePct: number;
  printWasteKg: number;
  printWastePct: number;
  machineWasteKg: number;
  machineWastePct: number;
  coverPatchWasteKg: number;
  coverPatchWastePct: number;
  totalWasteKg: number;
  totalWastePct: number;
}

export interface ValvomaticSummaryResult {
  qualities: ValvomaticQualitySummaryItem[];
  wastage: ValvomaticWastageBreakdown;
  overall: {
    totalReports: number;
    totalRolls: number;
    totalRollMtr: number;
    totalNetWt: number;
    avgGsm: number;
    totalProductionPcs: number;
    totalProductionKg: number;
    totalCoverPatchOs: number;
    totalCoverPatchDs: number;
    totalValvePatch: number;
    totalWastageKg: number;
    totalWastagePct: number;
    totalNetProductionKg: number;
  };
}